+++ lib/rspec/matchers/dsl.rb
@@ -4,9 +4,9 @@ module RSpec
       # Defines a custom matcher.
       # @see RSpec::Matchers
       def define(name, &declarations)
-        matcher_template = RSpec::Matchers::DSL::Matcher.new(name, &declarations)
+        matcher_template = RSpec::Matchers::DSL::Matcher.subclass(name, &declarations)
         define_method name do |*expected|
-          matcher = matcher_template.for_expected(*expected)
+          matcher = matcher_template.new(*expected)
           matcher.matcher_execution_context = @matcher_execution_context ||= self
           matcher
         end