+++ lib/rspec/matchers/matcher.rb
@@ -3,72 +3,16 @@ require 'set'
 module RSpec
   module Matchers
     module DSL
-      # Provides the context in which the block passed to RSpec::Matchers.define
-      # will be evaluated.
-      class Matcher
-        include RSpec::Matchers::Pretty
-        include RSpec::Matchers
-
-        attr_reader :expected, :actual, :rescued_exception
-        attr_accessor :matcher_execution_context
-
-        # @api private
-        def initialize(name, &declarations)
-          @name         = name
-          @declarations = declarations
-          @actual       = nil
-          @diffable     = false
-          @expected_exception, @rescued_exception = nil, nil
-          @match_for_should_not_block = nil
-          @messages = {}
-        end
-
-        PERSISTENT_INSTANCE_VARIABLES = [
-          :@name, :@declarations, :@diffable,
-          :@match_block, :@match_for_should_not_block,
-          :@expected_exception
-        ].to_set
-
-        # @api private
-        def for_expected(*expected)
-          @expected = expected
-          dup.instance_exec do
-            instance_variables.map {|ivar| ivar.intern}.each do |ivar|
-              instance_variable_set(ivar, nil) unless (PERSISTENT_INSTANCE_VARIABLES + [:@expected]).include?(ivar)
-            end
-            @messages = {}
-            instance_exec(*@expected, &@declarations)
-            self
-          end
-        end
-
-        # @api private
-        # Used internally by +should+ and +should_not+.
-        def matches?(actual)
-          @actual = actual
-          if @expected_exception
-            begin
-              instance_exec(actual, &@match_block)
-              true
-            rescue @expected_exception => @rescued_exception
-              false
-            end
-          else
-            begin
-              instance_exec(actual, &@match_block)
-            rescue RSpec::Expectations::ExpectationNotMetError
-              false
-            end
-          end
-        end
-
+      # Contains the methods that are available from within the
+      # `RSpec::Matchers.define` DSL for creating custom matchers.
+      module Macros
         # Stores the block that is used to determine whether this matcher passes
         # or fails. The block should return a boolean value. When the matcher is
-        # passed to `should` and the block returns `true`, then the expectation
-        # passes. Similarly, when the matcher is passed to `should_not` and the
+        # passed to `expect(...).to` and the block returns `true`, then the expectation
+        # passes. Similarly, when the matcher is passed to `expect(...).not_to` and the
         # block returns `false`, then the expectation passes.
         #
-        # Use `match_for_should` when used in conjuntion with
+        # Use `match_for_should` when used in conjunction with
         # `match_for_should_not`.
         #
         # @example
@@ -79,26 +23,35 @@ module RSpec
         #       end
         #     end
         #
-        #     4.should be_even     # passes
-        #     3.should_not be_even # passes
-        #     3.should be_even     # fails
-        #     4.should_not be_even # fails
+        #     expect(4).to be_even     # passes
+        #     expect(3).not_to be_even # passes
+        #     expect(3).to be_even     # fails
+        #     expect(4).not_to be_even # fails
         #
-        # @yield [Object] actual the actual value (or receiver of should)
-        def match(&block)
-          @match_block = block
+        # @yield [Object] actual the actual value (i.e. the value wrapped by `expect`)
+        def match(&match_block)
+          define_user_override(:matches?, match_block) do |actual|
+            begin
+              @actual = actual
+              super(*actual_arg_for(match_block))
+            rescue RSpec::Expectations::ExpectationNotMetError
+              false
+            end
+          end
         end
+        alias match_for_should match
 
-        alias_method :match_for_should, :match
-
-        # Use this to define the block for a negative expectation (`should_not`)
+        # Use this to define the block for a negative expectation (`expect(...).not_to`)
         # when the positive and negative forms require different handling. This
         # is rarely necessary, but can be helpful, for example, when specifying
         # asynchronous processes that require different timeouts.
         #
-        # @yield [Object] actual the actual value (or receiver of should)
-        def match_for_should_not(&block)
-          @match_for_should_not_block = block
+        # @yield [Object] actual the actual value (i.e. the value wrapped by `expect`)
+        def match_for_should_not(&match_block)
+          define_user_override(:does_not_match?, match_block) do |actual|
+            @actual = actual
+            super(*actual_arg_for(match_block))
+          end
         end
 
         # Use this instead of `match` when the block will raise an exception
@@ -112,51 +65,59 @@ module RSpec
         #       end
         #     end
         #
-        #     email_validator.should accept_as_valid("person@company.com")
-        def match_unless_raises(exception=Exception, &block)
-          @expected_exception = exception
-          match(&block)
+        #     expect(email_validator).to accept_as_valid("person@company.com")
+        #
+        # @yield [Object] actual the actual object (i.e. the value wrapped by `expect`)
+        def match_unless_raises(expected_exception=Exception, &match_block)
+          define_user_override(:matches?, match_block) do |actual|
+            @actual = actual
+            begin
+              super(*actual_arg_for(match_block))
+            rescue expected_exception => @rescued_exception
+              false
+            else
+              true
+            end
+          end
         end
 
-        # Customize the failure messsage to use when this matcher is invoked with
-        # `should`. Only use this when the message generated by default doesn't
-        # suit your needs.
+        # Customizes the failure messsage to use when this matcher is
+        # asked to positively match. Only use this when the message
+        # generated by default doesn't suit your needs.
         #
         # @example
         #
         #     RSpec::Matchers.define :have_strength do |expected|
-        #       match { ... }
+        #       match { your_match_logic }
         #
         #       failure_message_for_should do |actual|
         #         "Expected strength of #{expected}, but had #{actual.strength}"
         #       end
         #     end
         #
-        # @yield [Object] actual the actual object
-        def failure_message_for_should(&block)
-          cache_or_call_cached(:failure_message_for_should, &block)
+        # @yield [Object] actual the actual object (i.e. the value wrapped by `expect`)
+        def failure_message_for_should(&definition)
+          define_user_override(__method__, definition)
         end
 
-        # Customize the failure messsage to use when this matcher is invoked with
-        # `should_not`. Only use this when the message generated by default
-        # doesn't suit your needs.
+        # Customize the failure messsage to use when this matcher is asked
+        # to negatively match. Only use this when the message generated by
+        # default doesn't suit your needs.
         #
         # @example
         #
         #     RSpec::Matchers.define :have_strength do |expected|
-        #       match { ... }
+        #       match { your_match_logic }
         #
         #       failure_message_for_should_not do |actual|
         #         "Expected not to have strength of #{expected}, but did"
         #       end
         #     end
         #
-        # @yield [Object] actual the actual object
-        # @yield [Object] actual the actual object
-        def failure_message_for_should_not(&block)
-          cache_or_call_cached(:failure_message_for_should_not, &block)
+        # @yield [Object] actual the actual object (i.e. the value wrapped by `expect`)
+        def failure_message_for_should_not(&definition)
+          define_user_override(__method__, definition)
         end
-
 
         # Customize the description to use for one-liners.  Only use this when
         # the description generated by default doesn't suit your needs.
@@ -164,20 +125,22 @@ module RSpec
         # @example
         #
         #     RSpec::Matchers.define :qualify_for do |expected|
-        #       match { ... }
+        #       match { your_match_logic }
         #
         #       description do
         #         "qualify for #{expected}"
         #       end
         #     end
-        def description(&block)
-          cache_or_call_cached(:description, &block)
+        #
+        # @yield [Object] actual the actual object (i.e. the value wrapped by `expect`)
+        def description(&definition)
+          define_user_override(__method__, definition)
         end
 
         # Tells the matcher to diff the actual and expected values in the failure
         # message.
         def diffable
-          @diffable = true
+          define_method(:diffable?) { true }
         end
 
         # Convenience for defining methods on this matcher to create a fluent
@@ -197,35 +160,131 @@ module RSpec
         #       end
         #     end
         #
-        #     minor.should have_errors_on(:age).with("Not old enough to participate")
-        def chain(method, &block)
-          define_method method do |*args|
-            block.call(*args)
+        #     expect(minor).to have_errors_on(:age).with("Not old enough to participate")
+        def chain(name, &definition)
+          define_user_override(name, definition) do |*args, &block|
+            super(*args, &block)
             self
           end
         end
 
+      private
+
+        # Does the following:
+        #
+        # - Defines the named method usign a user-provided block
+        #   in self::UserMethodDefs, which is included as an ancestor
+        #   in the singleton class in which we eval the `define` block.
+        # - Defines an overriden definition for the same method
+        #   usign the provided `our_def` block.
+        # - Provides a default `our_def` block for the common case
+        #   of needing to call the user's definition with `@actual`
+        #   as an arg, but only if their block's arity can handle it.
+        #
+        # This compiles the user block into an actual method, allowing
+        # them to use normal method constructs like `return`
+        # (e.g. for a early guard statement), while allowing us to define
+        # an override that can provide the wrapped handling
+        # (e.g. assigning `@actual`, rescueing errors, etc) and
+        # can `super` to the user's definition.
+        def define_user_override(method_name, user_def, &our_def)
+          self::UserMethodDefs.__send__(:define_method, method_name, &user_def)
+          our_def ||= lambda { super(*actual_arg_for(user_def)) }
+          define_method(method_name, &our_def)
+        end
+      end
+
+      # Defines default implementations of the matcher
+      # protocol methods for custom matchers. You can
+      # override any of these using the {RSpec::Matchers::DSL::Macros Macros} methods
+      # from within an `RSpec::Matchers.define` block.
+      module DefaultImplementations
         # @api private
-        # Used internally by objects returns by +should+ and +should_not+.
+        # Used internally by objects returns by `should` and `should_not`.
         def diffable?
-          @diffable
+          false
         end
+
+        # The default description.
+        def description
+          "#{name_to_sentence}#{expected_to_sentence}"
+        end
+
+        # The default failure message for positive expectations.
+        def failure_message_for_should
+          "expected #{actual.inspect} to #{name_to_sentence}#{expected_to_sentence}"
+        end
+
+        # The default failure message for negative expectations.
+        def failure_message_for_should_not
+          "expected #{actual.inspect} not to #{name_to_sentence}#{expected_to_sentence}"
+        end
+      end
+
+      # Provides the base class for custom matchers. The block passed to
+      # `RSpec::Matchers.define` will be evaluated in the context of a subclass,
+      # and will have the {RSpec::Matchers::DSL::Macros Macros} methods available.
+      class Matcher
+        # Provides default implementations for the matcher protocol methods.
+        include DefaultImplementations
+
+        # Allows expectation expressions to be used in the match block.
+        include RSpec::Matchers
+
+        # Converts matcher name and expected args to an English expresion.
+        include RSpec::Matchers::Pretty
+
+        # Makes the macro methods available to an `RSpec::Matchers.define` block.
+        extend Macros
+
+        attr_reader   :expected, :actual, :rescued_exception
+        attr_accessor :matcher_execution_context
 
         # @api private
-        # Used internally by +should_not+
-        def does_not_match?(actual)
-          @actual = actual
-          @match_for_should_not_block ?
-            instance_exec(actual, &@match_for_should_not_block) :
-            !matches?(actual)
+        def initialize(*expected)
+          @actual   = nil
+          @expected = expected
+
+          class << self
+            # See `Macros#define_user_override` above, for an explanation.
+            include const_set(:UserMethodDefs, Module.new)
+            self
+          end.class_exec(*expected, &self.class.instance_variable_get(:@declarations))
         end
 
-        def respond_to?(method, include_private=false)
-          super || matcher_execution_context.respond_to?(method, include_private)
+        # Adds the expected value so we can identify an instance of
+        # the matcher in error messages (e.g. for `NoMethodError`)
+        def inspect
+          "#<#{self.class.name} expected=#{expected}>"
         end
 
-        private
+        if RUBY_VERSION.to_f >= 1.9
+          # Indicates that this matcher responds to messages
+          # from the `matcher_execution_context` as well.
+          # Also, supports getting a method object for such methods.
+          def respond_to_missing?(method, include_private=false)
+            super || matcher_execution_context.respond_to?(method, include_private)
+          end
+        else # for 1.8.7
+          # Indicates that this matcher responds to messages
+          # from the `matcher_execution_context` as well.
+          def respond_to?(method, include_private=false)
+            super || matcher_execution_context.respond_to?(method, include_private)
+          end
+        end
 
+      private
+
+        def actual_arg_for(block)
+          block.arity.zero? ? [] : [@actual]
+        end
+
+        # Takes care of forwarding unhandled messages to the
+        # `matcher_execution_context` (typically the current
+        # running `RSpec::Core::Example`). This is needed by
+        # rspec-rails so that it can define matchers that wrap
+        # Rails' test helper methods, but it's also a useful
+        # feature in its own right.
         def method_missing(method, *args, &block)
           if matcher_execution_context.respond_to?(method)
             matcher_execution_context.__send__ method, *args, &block
@@ -234,48 +293,32 @@ module RSpec
           end
         end
 
-        def include(*args)
-          singleton_class.__send__(:include, *args)
-        end
+        # @api private
+        def self.subclass(name, &declarations)
+          Class.new(self) do
+            define_method(:name) { name }
+            @declarations = declarations
+          end.tap do |klass|
+            const_name = ('_' + name.to_s).gsub(/_+([[:alpha:]])/) { $1.upcase }
 
-        def define_method(name, &block)
-          singleton_class.__send__(:define_method, name, &block)
-        end
+            const_name.gsub!(/\A_+/, '')             # Remove any leading underscores
+            const_name.gsub!(/\W/, '')               # JRuby, RBX and others don't like non-ascii in const names
+            const_name.gsub!(/\A([^A-Z]|\z)/, 'A\1') # Prepend a valid first letter if needed (A-Z)
 
-        def cache_or_call_cached(key, &block)
-          block ? cache(key, &block) : call_cached(key)
-        end
+            # Add a trailing number if needed to disambiguate from an existing constant.
+            if Custom.const_defined?(const_name)
+              const_name << "2"
+              const_name.next! while Custom.const_defined?(const_name)
+            end
 
-        def cache(key, &block)
-          @messages[key] = block
-        end
-
-        def call_cached(key)
-          if @messages.has_key?(key)
-            @messages[key].arity == 1 ? @messages[key].call(@actual) : @messages[key].call
-          else
-            __send__("default_#{key}")
-          end
-        end
-
-        def default_description
-          "#{name_to_sentence}#{expected_to_sentence}"
-        end
-
-        def default_failure_message_for_should
-          "expected #{actual.inspect} to #{name_to_sentence}#{expected_to_sentence}"
-        end
-
-        def default_failure_message_for_should_not
-          "expected #{actual.inspect} not to #{name_to_sentence}#{expected_to_sentence}"
-        end
-
-        unless method_defined?(:singleton_class)
-          def singleton_class
-            class << self; self; end
+            Custom.const_set(const_name, klass)
           end
         end
       end
+    end
+
+    # Namespace module that holds custom matcher classes.
+    module Custom
     end
   end
 end